chatbot-app/backend/data/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.4
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0